backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
python-dotenv
openpyxl
google-generativeai
sendgrid
